+++ Matrix.h
@@ -3,8 +3,8 @@
 class Matrix
 {
 public:
-	int rows = 0;
-	int cols = 0;
+	size_t rows = 0;
+	size_t cols = 0;
 
 	Matrix() = default;
 	Matrix(const Matrix&) = delete;
@@ -12,8 +12,8 @@ public:
 	Matrix& operator=(const Matrix& other) = default;
 	Matrix(Matrix&&) = default;
 
-	virtual int Size();
-	[[nodiscard]] virtual int Size() const;
+	virtual size_t Size();
+	[[nodiscard]] virtual size_t Size() const;
 	virtual void Fill(long double value) = 0;
 
 	virtual ~Matrix() = default;
